fix: select vocab & synonyms directly from review instead of current
if a word is the only remaining item in quiz and we add a synonym, it isn't added to quiz>current so
for matching answers we should always use the definitely updated/accurate review entry. However in
most other stats (correct/streak) etc we should continue to use current for those purposes.